requirements.txt
argparse
opencv-python
guibot
pyautogui
pyyaml